* EZDistribution.m ([EZDistribution -output]): Use putDouble 	instead of putInt.
	* EZBin.m ([EZBin -setAxisLabelsX:Y:],
	[EZBin -setPrecision:]): Put these messages in Creating phase.


git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@7198 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/analysis/EZDistribution.m
@@ -55,11 +55,11 @@ PHASE(Using)
 
   if (fileOutput)
     {
-      [anOutFile putInt: probabilities[0]];
+      [anOutFile putDouble: probabilities[0]];
       for (i = 1; i < binCount; i++)
         {
           [anOutFile putTab];
-          [anOutFile putInt: probabilities[i]];
+          [anOutFile putDouble: probabilities[i]];
         }
       [anOutFile putNewLine];
     }
